@@ -8,10 +8,13 @@ public class AbstractVisitor<R,A> implements AllVisitor<R,A> {
     /* Expr */
     public R visit(hardtyped.Absyn.Function p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.ApplyFunction p, A arg) { return visitDefault(p, arg); }
+    public R visit(hardtyped.Absyn.PrintFunction p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.Sum p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.Substract p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.Multiply p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.Divide p, A arg) { return visitDefault(p, arg); }
+    public R visit(hardtyped.Absyn.UnaryPlus p, A arg) { return visitDefault(p, arg); }
+    public R visit(hardtyped.Absyn.UnaryMinus p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.Let p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.LetInference p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.LetAscription p, A arg) { return visitDefault(p, arg); }
@@ -19,6 +22,7 @@ public class AbstractVisitor<R,A> implements AllVisitor<R,A> {
     public R visit(hardtyped.Absyn.RealValue p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.StringValue p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.AtomicExpression p, A arg) { return visitDefault(p, arg); }
+    public R visit(hardtyped.Absyn.ParenthesesExpression p, A arg) { return visitDefault(p, arg); }
     public R visitDefault(hardtyped.Absyn.Expr p, A arg) {
       throw new IllegalArgumentException(this.getClass().getName() + ": " + p);
     }
@@ -29,9 +33,13 @@ public class AbstractVisitor<R,A> implements AllVisitor<R,A> {
     public R visitDefault(hardtyped.Absyn.InExpr p, A arg) {
       throw new IllegalArgumentException(this.getClass().getName() + ": " + p);
     }
+    /* FuncArg */
+    public R visit(hardtyped.Absyn.MultipleArgs p, A arg) { return visitDefault(p, arg); }
+    public R visit(hardtyped.Absyn.FinalArg p, A arg) { return visitDefault(p, arg); }
+    public R visitDefault(hardtyped.Absyn.FuncArg p, A arg) {
+      throw new IllegalArgumentException(this.getClass().getName() + ": " + p);
+    }
     /* VarDec */
-    public R visit(hardtyped.Absyn.MultipleVars p, A arg) { return visitDefault(p, arg); }
-    public R visit(hardtyped.Absyn.MultipleVarsFinal p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.TypedVar p, A arg) { return visitDefault(p, arg); }
     public R visit(hardtyped.Absyn.UntypedVar p, A arg) { return visitDefault(p, arg); }
     public R visitDefault(hardtyped.Absyn.VarDec p, A arg) {
