@@ -7,21 +7,28 @@ package hardtyped;
 public class ComposVisitor<A> implements
   hardtyped.Absyn.Expr.Visitor<hardtyped.Absyn.Expr,A>,
   hardtyped.Absyn.InExpr.Visitor<hardtyped.Absyn.InExpr,A>,
+  hardtyped.Absyn.FuncArg.Visitor<hardtyped.Absyn.FuncArg,A>,
   hardtyped.Absyn.VarDec.Visitor<hardtyped.Absyn.VarDec,A>,
   hardtyped.Absyn.Type.Visitor<hardtyped.Absyn.Type,A>
 {
     /* Expr */
     public hardtyped.Absyn.Expr visit(hardtyped.Absyn.Function p, A arg)
     {
-      hardtyped.Absyn.VarDec vardec_ = p.vardec_.accept(this, arg);
+      hardtyped.Absyn.FuncArg funcarg_ = p.funcarg_.accept(this, arg);
       hardtyped.Absyn.InExpr inexpr_ = p.inexpr_.accept(this, arg);
-      return new hardtyped.Absyn.Function(vardec_, inexpr_);
+      return new hardtyped.Absyn.Function(funcarg_, inexpr_);
     }
     public hardtyped.Absyn.Expr visit(hardtyped.Absyn.ApplyFunction p, A arg)
     {
       hardtyped.Absyn.Expr expr_1 = p.expr_1.accept(this, arg);
       hardtyped.Absyn.Expr expr_2 = p.expr_2.accept(this, arg);
       return new hardtyped.Absyn.ApplyFunction(expr_1, expr_2);
+    }
+    public hardtyped.Absyn.Expr visit(hardtyped.Absyn.PrintFunction p, A arg)
+    {
+      String print_ = p.print_;
+      hardtyped.Absyn.Expr expr_ = p.expr_.accept(this, arg);
+      return new hardtyped.Absyn.PrintFunction(print_, expr_);
     }
     public hardtyped.Absyn.Expr visit(hardtyped.Absyn.Sum p, A arg)
     {
@@ -46,6 +53,16 @@ public class ComposVisitor<A> implements
       hardtyped.Absyn.Expr expr_1 = p.expr_1.accept(this, arg);
       hardtyped.Absyn.Expr expr_2 = p.expr_2.accept(this, arg);
       return new hardtyped.Absyn.Divide(expr_1, expr_2);
+    }
+    public hardtyped.Absyn.Expr visit(hardtyped.Absyn.UnaryPlus p, A arg)
+    {
+      hardtyped.Absyn.Expr expr_ = p.expr_.accept(this, arg);
+      return new hardtyped.Absyn.UnaryPlus(expr_);
+    }
+    public hardtyped.Absyn.Expr visit(hardtyped.Absyn.UnaryMinus p, A arg)
+    {
+      hardtyped.Absyn.Expr expr_ = p.expr_.accept(this, arg);
+      return new hardtyped.Absyn.UnaryMinus(expr_);
     }
     public hardtyped.Absyn.Expr visit(hardtyped.Absyn.Let p, A arg)
     {
@@ -87,6 +104,11 @@ public class ComposVisitor<A> implements
       String ident_ = p.ident_;
       return new hardtyped.Absyn.AtomicExpression(ident_);
     }
+    public hardtyped.Absyn.Expr visit(hardtyped.Absyn.ParenthesesExpression p, A arg)
+    {
+      hardtyped.Absyn.Expr expr_ = p.expr_.accept(this, arg);
+      return new hardtyped.Absyn.ParenthesesExpression(expr_);
+    }
 
     /* InExpr */
     public hardtyped.Absyn.InExpr visit(hardtyped.Absyn.MultipleExpressions p, A arg)
@@ -106,18 +128,20 @@ public class ComposVisitor<A> implements
       return new hardtyped.Absyn.InnerExpression(expr_);
     }
 
-    /* VarDec */
-    public hardtyped.Absyn.VarDec visit(hardtyped.Absyn.MultipleVars p, A arg)
-    {
-      hardtyped.Absyn.VarDec vardec_1 = p.vardec_1.accept(this, arg);
-      hardtyped.Absyn.VarDec vardec_2 = p.vardec_2.accept(this, arg);
-      return new hardtyped.Absyn.MultipleVars(vardec_1, vardec_2);
-    }
-    public hardtyped.Absyn.VarDec visit(hardtyped.Absyn.MultipleVarsFinal p, A arg)
+    /* FuncArg */
+    public hardtyped.Absyn.FuncArg visit(hardtyped.Absyn.MultipleArgs p, A arg)
     {
       hardtyped.Absyn.VarDec vardec_ = p.vardec_.accept(this, arg);
-      return new hardtyped.Absyn.MultipleVarsFinal(vardec_);
+      hardtyped.Absyn.FuncArg funcarg_ = p.funcarg_.accept(this, arg);
+      return new hardtyped.Absyn.MultipleArgs(vardec_, funcarg_);
     }
+    public hardtyped.Absyn.FuncArg visit(hardtyped.Absyn.FinalArg p, A arg)
+    {
+      hardtyped.Absyn.VarDec vardec_ = p.vardec_.accept(this, arg);
+      return new hardtyped.Absyn.FinalArg(vardec_);
+    }
+
+    /* VarDec */
     public hardtyped.Absyn.VarDec visit(hardtyped.Absyn.TypedVar p, A arg)
     {
       String ident_ = p.ident_;
