@@ -10,7 +10,7 @@ public abstract class FoldVisitor<R,A> implements AllVisitor<R,A> {
 /* Expr */
     public R visit(hardtyped.Absyn.Function p, A arg) {
       R r = leaf(arg);
-      r = combine(p.vardec_.accept(this, arg), r, arg);
+      r = combine(p.funcarg_.accept(this, arg), r, arg);
       r = combine(p.inexpr_.accept(this, arg), r, arg);
       return r;
     }
@@ -18,6 +18,11 @@ public abstract class FoldVisitor<R,A> implements AllVisitor<R,A> {
       R r = leaf(arg);
       r = combine(p.expr_1.accept(this, arg), r, arg);
       r = combine(p.expr_2.accept(this, arg), r, arg);
+      return r;
+    }
+    public R visit(hardtyped.Absyn.PrintFunction p, A arg) {
+      R r = leaf(arg);
+      r = combine(p.expr_.accept(this, arg), r, arg);
       return r;
     }
     public R visit(hardtyped.Absyn.Sum p, A arg) {
@@ -42,6 +47,16 @@ public abstract class FoldVisitor<R,A> implements AllVisitor<R,A> {
       R r = leaf(arg);
       r = combine(p.expr_1.accept(this, arg), r, arg);
       r = combine(p.expr_2.accept(this, arg), r, arg);
+      return r;
+    }
+    public R visit(hardtyped.Absyn.UnaryPlus p, A arg) {
+      R r = leaf(arg);
+      r = combine(p.expr_.accept(this, arg), r, arg);
+      return r;
+    }
+    public R visit(hardtyped.Absyn.UnaryMinus p, A arg) {
+      R r = leaf(arg);
+      r = combine(p.expr_.accept(this, arg), r, arg);
       return r;
     }
     public R visit(hardtyped.Absyn.Let p, A arg) {
@@ -80,6 +95,11 @@ public abstract class FoldVisitor<R,A> implements AllVisitor<R,A> {
       R r = leaf(arg);
       return r;
     }
+    public R visit(hardtyped.Absyn.ParenthesesExpression p, A arg) {
+      R r = leaf(arg);
+      r = combine(p.expr_.accept(this, arg), r, arg);
+      return r;
+    }
 
 /* InExpr */
     public R visit(hardtyped.Absyn.MultipleExpressions p, A arg) {
@@ -99,18 +119,20 @@ public abstract class FoldVisitor<R,A> implements AllVisitor<R,A> {
       return r;
     }
 
-/* VarDec */
-    public R visit(hardtyped.Absyn.MultipleVars p, A arg) {
+/* FuncArg */
+    public R visit(hardtyped.Absyn.MultipleArgs p, A arg) {
       R r = leaf(arg);
-      r = combine(p.vardec_1.accept(this, arg), r, arg);
-      r = combine(p.vardec_2.accept(this, arg), r, arg);
+      r = combine(p.vardec_.accept(this, arg), r, arg);
+      r = combine(p.funcarg_.accept(this, arg), r, arg);
       return r;
     }
-    public R visit(hardtyped.Absyn.MultipleVarsFinal p, A arg) {
+    public R visit(hardtyped.Absyn.FinalArg p, A arg) {
       R r = leaf(arg);
       r = combine(p.vardec_.accept(this, arg), r, arg);
       return r;
     }
+
+/* VarDec */
     public R visit(hardtyped.Absyn.TypedVar p, A arg) {
       R r = leaf(arg);
       r = combine(p.type_.accept(this, arg), r, arg);
