@@ -110,6 +110,21 @@ public class PrettyPrinter
     buf_.delete(0,buf_.length());
     return temp;
   }
+  public static String print(hardtyped.Absyn.FuncArg foo)
+  {
+    pp(foo, 0);
+    trim();
+    String temp = buf_.toString();
+    buf_.delete(0,buf_.length());
+    return temp;
+  }
+  public static String show(hardtyped.Absyn.FuncArg foo)
+  {
+    sh(foo);
+    String temp = buf_.toString();
+    buf_.delete(0,buf_.length());
+    return temp;
+  }
   public static String print(hardtyped.Absyn.VarDec foo)
   {
     pp(foo, 0);
@@ -171,7 +186,7 @@ public class PrettyPrinter
        hardtyped.Absyn.Function _function = (hardtyped.Absyn.Function) foo;
        if (_i_ > 0) render(_L_PAREN);
        render("/\\");
-       pp(_function.vardec_, 0);
+       pp(_function.funcarg_, 0);
        render("{");
        pp(_function.inexpr_, 0);
        render("}");
@@ -184,6 +199,16 @@ public class PrettyPrinter
        pp(_applyfunction.expr_1, 0);
        render("(");
        pp(_applyfunction.expr_2, 0);
+       render(")");
+       if (_i_ > 0) render(_R_PAREN);
+    }
+    else     if (foo instanceof hardtyped.Absyn.PrintFunction)
+    {
+       hardtyped.Absyn.PrintFunction _printfunction = (hardtyped.Absyn.PrintFunction) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       pp(_printfunction.print_, 0);
+       render("(");
+       pp(_printfunction.expr_, 0);
        render(")");
        if (_i_ > 0) render(_R_PAREN);
     }
@@ -221,6 +246,22 @@ public class PrettyPrinter
        pp(_divide.expr_1, 0);
        render("/");
        pp(_divide.expr_2, 0);
+       if (_i_ > 0) render(_R_PAREN);
+    }
+    else     if (foo instanceof hardtyped.Absyn.UnaryPlus)
+    {
+       hardtyped.Absyn.UnaryPlus _unaryplus = (hardtyped.Absyn.UnaryPlus) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       render("+");
+       pp(_unaryplus.expr_, 0);
+       if (_i_ > 0) render(_R_PAREN);
+    }
+    else     if (foo instanceof hardtyped.Absyn.UnaryMinus)
+    {
+       hardtyped.Absyn.UnaryMinus _unaryminus = (hardtyped.Absyn.UnaryMinus) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       render("-");
+       pp(_unaryminus.expr_, 0);
        if (_i_ > 0) render(_R_PAREN);
     }
     else     if (foo instanceof hardtyped.Absyn.Let)
@@ -285,6 +326,15 @@ public class PrettyPrinter
        pp(_atomicexpression.ident_, 0);
        if (_i_ > 0) render(_R_PAREN);
     }
+    else     if (foo instanceof hardtyped.Absyn.ParenthesesExpression)
+    {
+       hardtyped.Absyn.ParenthesesExpression _parenthesesexpression = (hardtyped.Absyn.ParenthesesExpression) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       render("(");
+       pp(_parenthesesexpression.expr_, 0);
+       render(")");
+       if (_i_ > 0) render(_R_PAREN);
+    }
 
   }
 
@@ -317,26 +367,31 @@ public class PrettyPrinter
 
   }
 
+  private static void pp(hardtyped.Absyn.FuncArg foo, int _i_)
+  {
+    if (foo instanceof hardtyped.Absyn.MultipleArgs)
+    {
+       hardtyped.Absyn.MultipleArgs _multipleargs = (hardtyped.Absyn.MultipleArgs) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       pp(_multipleargs.vardec_, 0);
+       render(".");
+       pp(_multipleargs.funcarg_, 0);
+       if (_i_ > 0) render(_R_PAREN);
+    }
+    else     if (foo instanceof hardtyped.Absyn.FinalArg)
+    {
+       hardtyped.Absyn.FinalArg _finalarg = (hardtyped.Absyn.FinalArg) foo;
+       if (_i_ > 0) render(_L_PAREN);
+       pp(_finalarg.vardec_, 0);
+       render(".");
+       if (_i_ > 0) render(_R_PAREN);
+    }
+
+  }
+
   private static void pp(hardtyped.Absyn.VarDec foo, int _i_)
   {
-    if (foo instanceof hardtyped.Absyn.MultipleVars)
-    {
-       hardtyped.Absyn.MultipleVars _multiplevars = (hardtyped.Absyn.MultipleVars) foo;
-       if (_i_ > 0) render(_L_PAREN);
-       pp(_multiplevars.vardec_1, 0);
-       render(".");
-       pp(_multiplevars.vardec_2, 0);
-       if (_i_ > 0) render(_R_PAREN);
-    }
-    else     if (foo instanceof hardtyped.Absyn.MultipleVarsFinal)
-    {
-       hardtyped.Absyn.MultipleVarsFinal _multiplevarsfinal = (hardtyped.Absyn.MultipleVarsFinal) foo;
-       if (_i_ > 0) render(_L_PAREN);
-       pp(_multiplevarsfinal.vardec_, 0);
-       render(".");
-       if (_i_ > 0) render(_R_PAREN);
-    }
-    else     if (foo instanceof hardtyped.Absyn.TypedVar)
+    if (foo instanceof hardtyped.Absyn.TypedVar)
     {
        hardtyped.Absyn.TypedVar _typedvar = (hardtyped.Absyn.TypedVar) foo;
        if (_i_ > 0) render(_L_PAREN);
@@ -413,7 +468,7 @@ public class PrettyPrinter
        hardtyped.Absyn.Function _function = (hardtyped.Absyn.Function) foo;
        render("(");
        render("Function");
-       sh(_function.vardec_);
+       sh(_function.funcarg_);
        sh(_function.inexpr_);
        render(")");
     }
@@ -424,6 +479,15 @@ public class PrettyPrinter
        render("ApplyFunction");
        sh(_applyfunction.expr_1);
        sh(_applyfunction.expr_2);
+       render(")");
+    }
+    if (foo instanceof hardtyped.Absyn.PrintFunction)
+    {
+       hardtyped.Absyn.PrintFunction _printfunction = (hardtyped.Absyn.PrintFunction) foo;
+       render("(");
+       render("PrintFunction");
+       sh(_printfunction.print_);
+       sh(_printfunction.expr_);
        render(")");
     }
     if (foo instanceof hardtyped.Absyn.Sum)
@@ -460,6 +524,22 @@ public class PrettyPrinter
        render("Divide");
        sh(_divide.expr_1);
        sh(_divide.expr_2);
+       render(")");
+    }
+    if (foo instanceof hardtyped.Absyn.UnaryPlus)
+    {
+       hardtyped.Absyn.UnaryPlus _unaryplus = (hardtyped.Absyn.UnaryPlus) foo;
+       render("(");
+       render("UnaryPlus");
+       sh(_unaryplus.expr_);
+       render(")");
+    }
+    if (foo instanceof hardtyped.Absyn.UnaryMinus)
+    {
+       hardtyped.Absyn.UnaryMinus _unaryminus = (hardtyped.Absyn.UnaryMinus) foo;
+       render("(");
+       render("UnaryMinus");
+       sh(_unaryminus.expr_);
        render(")");
     }
     if (foo instanceof hardtyped.Absyn.Let)
@@ -523,6 +603,14 @@ public class PrettyPrinter
        sh(_atomicexpression.ident_);
        render(")");
     }
+    if (foo instanceof hardtyped.Absyn.ParenthesesExpression)
+    {
+       hardtyped.Absyn.ParenthesesExpression _parenthesesexpression = (hardtyped.Absyn.ParenthesesExpression) foo;
+       render("(");
+       render("ParenthesesExpression");
+       sh(_parenthesesexpression.expr_);
+       render(")");
+    }
   }
 
   private static void sh(hardtyped.Absyn.InExpr foo)
@@ -554,25 +642,29 @@ public class PrettyPrinter
     }
   }
 
+  private static void sh(hardtyped.Absyn.FuncArg foo)
+  {
+    if (foo instanceof hardtyped.Absyn.MultipleArgs)
+    {
+       hardtyped.Absyn.MultipleArgs _multipleargs = (hardtyped.Absyn.MultipleArgs) foo;
+       render("(");
+       render("MultipleArgs");
+       sh(_multipleargs.vardec_);
+       sh(_multipleargs.funcarg_);
+       render(")");
+    }
+    if (foo instanceof hardtyped.Absyn.FinalArg)
+    {
+       hardtyped.Absyn.FinalArg _finalarg = (hardtyped.Absyn.FinalArg) foo;
+       render("(");
+       render("FinalArg");
+       sh(_finalarg.vardec_);
+       render(")");
+    }
+  }
+
   private static void sh(hardtyped.Absyn.VarDec foo)
   {
-    if (foo instanceof hardtyped.Absyn.MultipleVars)
-    {
-       hardtyped.Absyn.MultipleVars _multiplevars = (hardtyped.Absyn.MultipleVars) foo;
-       render("(");
-       render("MultipleVars");
-       sh(_multiplevars.vardec_1);
-       sh(_multiplevars.vardec_2);
-       render(")");
-    }
-    if (foo instanceof hardtyped.Absyn.MultipleVarsFinal)
-    {
-       hardtyped.Absyn.MultipleVarsFinal _multiplevarsfinal = (hardtyped.Absyn.MultipleVarsFinal) foo;
-       render("(");
-       render("MultipleVarsFinal");
-       sh(_multiplevarsfinal.vardec_);
-       render(")");
-    }
     if (foo instanceof hardtyped.Absyn.TypedVar)
     {
        hardtyped.Absyn.TypedVar _typedvar = (hardtyped.Absyn.TypedVar) foo;
